backend/requirements.txt
google-cloud-firestore
flask<3.0,>=1.0
pylint
